Migrate testing from pytest-dockerc to using python-on-whales
The pytest-dockerc plug is unmaintained and there is now a dependency
issues with PyYAML because of the release of Cython v3 (which is itself
a build dependency for PyYAML). After some research this seemed like
the most similar package in terms of functionality to the package we
are replacing. Although it is not a pytest plugin it still provides
similar access and uses the Docker composition defined in the
repository.